// api/user/src/domains/customer/CustomerEnum.ts
export enum CustomerType {
    REGULAR = 'REGULAR',
    GUEST = 'GUEST',
}

export enum GenderType {
    MALE = 'MALE',
    FEMALE = 'FEMALE',
    OTHER = 'OTHER',
}

export enum ServiceType {
    CoWorkingSpace = 'CoWorkingSpace',
    Reservation = 'Reservation',
    Application = 'Application',
}

export enum SocialAccount {
    FACEBOOK = 'FACEBOOK',
    TWITTER = 'TWITTER',
    APPLE = 'APPLE',
    GOOGLE = 'GOOGLE',
    EMAIL = 'EMAIL',
}